refactor: extract auth into tribal-auth crate
Move AuthenticatedPrincipal, AuthContext, AuthError, Authenticator,
TransportAuthStrategy, and the bearer-token middleware out of
tribal-mcp into a new tribal-auth workspace crate. tribal-mcp now
depends on tribal-auth for both validation and middleware;
tribal-server imports require_bearer_auth and the associated state
from tribal-auth directly. The integration test verify_token.rs
moves with the code.

Scope, AuthToken, AuthTokenRepository, PgAuthTokenRepository, the
auth_tokens migrations, and the tribal token CLI lifecycle stay
where they are; only the runtime-bearing auth code moves.

Adds the subtle crate to the workspace, reserved for constant-time
cryptographic comparisons in the OAuth flow added in subsequent
commits.
