+++ crates/tribal-auth/src/lib.rs
@@ -1,0 +1,27 @@
+#![warn(clippy::pedantic)]
+#![deny(warnings)]
+//! Authentication and authorisation for Tribal.
+//!
+//! Owns the bearer-token resource-server plane (validation, scope
+//! enforcement, middleware response shaping) and the OAuth 2.1
+//! authorisation-server endpoints (CIMD discovery, DCR, /authorize,
+//! /token, well-known metadata). Token issuance from any path
+//! produces a row in the existing `auth_tokens` store that the
+//! bearer middleware verifies identically to a CLI-minted token.
+
+mod authenticator;
+mod context;
+mod error;
+mod middleware;
+mod principal;
+mod strategy;
+
+pub use authenticator::Authenticator;
+pub use context::AuthContext;
+pub use error::{
+    AuthError, DISPLAY_INVALID_TOKEN, DISPLAY_MISSING_TOKEN, DISPLAY_TOKEN_EXPIRED,
+    DISPLAY_TOKEN_REVOKED,
+};
+pub use middleware::{AuthMiddlewareState, require_bearer_auth};
+pub use principal::AuthenticatedPrincipal;
+pub use strategy::TransportAuthStrategy;
