+++ crates/tribal-mcp/src/lib.rs
@@ -4,14 +4,12 @@
 //! state management, and transport setup (stdio, HTTP, SSE).
 
 mod app_state;
-mod auth;
 mod config;
 mod error;
 mod fingerprint;
 mod format;
 mod handlers;
 mod mapping;
-mod middleware;
 mod polling;
 mod server_handler;
 mod session;
@@ -21,14 +19,9 @@ mod test_utils;
 mod tools;
 
 pub use app_state::{AppState, ResolvedProject};
-pub use auth::{
-    AuthContext, AuthError, AuthenticatedPrincipal, Authenticator, DISPLAY_INVALID_TOKEN,
-    DISPLAY_MISSING_TOKEN, DISPLAY_TOKEN_EXPIRED, DISPLAY_TOKEN_REVOKED, TransportAuthStrategy,
-};
 pub use config::{HandlerConfig, HandlerDiscoveryConfig, HandlerExplorationConfig};
 pub use error::{IntoCallToolResult, IntoMcpError, McpToolError};
 pub use fingerprint::build_inference_parameters;
-pub use middleware::{AuthMiddlewareState, require_bearer_auth};
 pub use server_handler::{ActivePromptVersions, ConnectionRepositories, TribalServerHandler};
 pub use session::{SessionActor, SessionContext, SessionProject};
 #[cfg(feature = "test-helpers")]
