+++ crates/tribal-server/Cargo.toml
@@ -59,6 +59,7 @@ tower.workspace = true
 tracing.workspace = true
 uuid.workspace = true
 
+tribal-auth.workspace = true
 tribal-common.workspace = true
 tribal-config.workspace = true
 tribal-db.workspace = true
