+++ crates/tribal-server/src/commands/check/checks/valid_token_exists.rs
@@ -12,7 +12,7 @@ use chrono::Utc;
 use sqlx::PgPool;
 use tribal_config::{Auth, CredentialsReadError, ENV_AUTH_TOKEN, TransportKind, read_credentials};
 use tribal_db::{AuthTokenRepository, PgAuthTokenRepository, PgPrincipalRepository};
-use tribal_mcp::{AuthError, Authenticator};
+use tribal_auth::{AuthError, Authenticator};
 
 use super::{
     state::CheckState,
